add more tests

+++ server/src/use-cases/feedbacks-use-cases/delete-feedback-use-case.ts
@@ -4,6 +4,10 @@ export class DeleteFeedbackUseCase {
   constructor(private feedbacksRepository: FeedbacksRepository) {}
 
   async execute(id: string) {
+    if (!id) {
+      throw new Error("Id is required");
+    }
+
     try {
       await this.feedbacksRepository.delete(id);
     } catch (error) {
